# resourceflow-cli/openapi/v1.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "ResourceFlowApi | v1",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost/"
    }
  ],
  "paths": {
...
    "/api/Admin/bookings/{id}": {
      "get": {
        "tags": [
          "Admin"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "Admin"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "put": {
        "tags": [
          "Admin"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminUpdateBookingRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminUpdateBookingRequest"
              }
            },
            "application/*\u002Bjson": {
              "schema": {
                "$ref": "#/components/schemas/AdminUpdateBookingRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/venues/{venueId}/bookings": {
      "get": {
        "tags": [
          "Bookings"
        ],
        "parameters": [
          {
            "name": "venueId",
            "in": "path",
            "required": true,
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Bookings/{id}": {
      "get": {
        "tags": [
          "Bookings"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "put": {
        "tags": [
          "Bookings"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BookingDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/BookingDto"
              }
            },
            "application/*\u002Bjson": {
              "schema": {
                "$ref": "#/components/schemas/BookingDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "Bookings"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Bookings/ref/{bookingRef}": {
      "get": {
        "tags": [
          "Bookings"
        ],
        "parameters": [
          {
            "name": "bookingRef",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "email",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Bookings": {
      "post": {
        "tags": [
          "Bookings"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BookingDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/BookingDto"
              }
            },
            "application/*\u002Bjson": {
              "schema": {
                "$ref": "#/components/schemas/BookingDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/Bookings/ref/{bookingRef}/cancel": {
      "post": {
        "tags": [
          "Bookings"
        ],
        "parameters": [
          {
            "name": "bookingRef",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CancelBookingByRefRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CancelBookingByRefRequest"
              }
            },
            "application/*\u002Bjson": {
              "schema": {
                "$ref": "#/components/schemas/CancelBookingByRefRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Bookings/ref/{bookingRef}/reminders": {
      "post": {
        "tags": [
          "Bookings"
        ],
        "parameters": [
          {
            "name": "bookingRef",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GuestReminderSubscribeRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/GuestReminderSubscribeRequest"
              }
            },
            "application/*\u002Bjson": {
              "schema": {
                "$ref": "#/components/schemas/GuestReminderSubscribeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "Bookings"
        ],
        "parameters": [
          {
            "name": "bookingRef",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GuestReminderUnsubscribeRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/GuestReminderUnsubscribeRequest"
              }
            },
            "application/*\u002Bjson": {
              "schema": {
                "$ref": "#/components/schemas/GuestReminderUnsubscribeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Bookings/ref/{bookingRef}/wallet/apple.pkpass": {
      "get": {
        "tags": [
          "Bookings"
        ],
        "parameters": [
          {
            "name": "bookingRef",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "email",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Bookings/ref/{bookingRef}/wallet/google": {
      "get": {
        "tags": [
          "Bookings"
        ],
        "parameters": [
          {
            "name": "bookingRef",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "email",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AdminUpdateBookingRequest": {
        "type": "object",
        "properties": {
          "venueId": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "sectionId": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "resourceId": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "date": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "partySize": {
            "maximum": 50,
            "minimum": 1,
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "customerEmail": {
            "type": [
              "null",
              "string"
            ]
          },
          "customerName": {
            "type": [
              "null",
              "string"
            ]
          },
          "specialRequests": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "BookingDto": {
        "type": "object",
        "properties": {
          "id": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "resourceId": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "sectionId": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "resourceGroupId": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "description": "When set, this booking reserves a combinable-resource group instead of a single resource.\nIReadOnlyList\u0026lt;int\u0026gt;? BookingDto.MemberResourceIds carries the group\u0027s member ids for the conflict check; both are\nresolved server-side. Only int? BookingDto.ResourceGroupId is persisted on the booking entity.",
            "format": "int32"
          },
          "memberResourceIds": {
            "type": [
              "null",
              "array"
            ],
            "items": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32"
            },
            "description": "Member resource ids the server resolved for int? BookingDto.ResourceGroupId while auto-assigning.\nNever read back from a client request \u2014 the booking path always re-resolves the members from\nthe persisted group, so a caller can\u0027t shrink this list to dodge a conflict check."
          },
          "venueId": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "date": {
            "type": "string",
            "format": "date-time"
          },
          "customerEmail": {
            "type": [
              "null",
              "string"
            ]
          },
          "customerName": {
            "type": [
              "null",
              "string"
            ]
          },
          "partySize": {
            "maximum": 50,
            "minimum": 1,
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "description": "Party size. Bounded to [int BookingLimits.MinPartySize, int BookingLimits.MaxPartySize] on create.",
            "format": "int32"
          },
          "isHeld": {
            "type": "boolean"
          },
          "specialRequests": {
            "type": [
              "null",
              "string"
            ]
          },
          "bookingRef": {
            "type": [
              "null",
              "string"
            ]
          },
          "endTime": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "resourceName": {
            "type": [
              "null",
              "string"
            ]
          },
          "sectionName": {
            "type": [
              "null",
              "string"
            ]
          },
          "resourceCapacity": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "isCancelled": {
            "type": "boolean"
          },
          "cancelledAt": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "holdId": {
            "type": [
              "null",
              "string"
            ],
            "description": "Optional hold ID obtained from POST /api/holds.\nProvide this when creating a booking to validate and consume the hold.\nNot returned in responses."
          }
        }
      },
...
      "CancelBookingByRefRequest": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          }
        }
      },
...
      "GuestReminderSubscribeRequest": {
        "required": [
          "email",
          "channel",
          "endpoint"
        ],
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          },
          "channel": {
            "maxLength": 16,
            "minLength": 0,
            "type": "string",
            "description": "\u0022expo\u0022 from the native app, \u0022webpush\u0022 from a browser."
          },
          "endpoint": {
            "maxLength": 2048,
            "minLength": 0,
            "type": "string",
            "description": "The Expo push token, or the Web Push subscription endpoint."
          },
          "p256dh": {
            "maxLength": 512,
            "minLength": 0,
            "type": [
              "null",
              "string"
            ]
          },
          "auth": {
            "maxLength": 512,
            "minLength": 0,
            "type": [
              "null",
              "string"
            ]
          },
          "locale": {
            "maxLength": 8,
            "minLength": 0,
            "type": [
              "null",
              "string"
            ]
          }
        },
        "description": "A guest asking to be reminded about a booking. \u0060Email\u0060 is the second half of the\nguest\u0027s identity alongside the reference in the route, exactly as on lookup and cancel."
      },
      "GuestReminderUnsubscribeRequest": {
        "required": [
          "email",
          "endpoint"
        ],
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          },
          "endpoint": {
            "maxLength": 2048,
            "minLength": 0,
            "type": "string"
          }
        }
      },
...
    }
  },
...
}